+++ src/yafowil/bootstrap/tests.py
@@ -1,30 +1,11 @@
-from interlude import interact
-from pprint import pprint
-from yafowil.tests import pxml
-import doctest
 import unittest
 
 
-optionflags = doctest.NORMALIZE_WHITESPACE | \
-              doctest.ELLIPSIS | \
-              doctest.REPORT_ONLY_FIRST_FAILURE
+class TestYafowilWebob(unittest.TestCase):
 
-TESTFILES = [
-    'common.rst',
-]
-
-
-def test_suite():
-    return unittest.TestSuite([
-        doctest.DocFileSuite(
-            file, 
-            optionflags=optionflags,
-            globs={'interact': interact,
-                   'pprint': pprint,
-                   'pxml': pxml},
-        ) for file in TESTFILES
-    ])
+    def test_dummy(self):
+        self.assertTrue(True)
 
 
 if __name__ == '__main__':
-    unittest.main(defaultTest='test_suite')                  #pragma NO COVER
+    unittest.main()